Web/requirements.txt
streamlit
python-dotenv
requests
graphviz

